RenderSystem changes and new methods
Made some modifications to improve the code within RenderSystem. For example the fullscreen if else block was uneeded. Also added some new functionality in the form of WindowBackgroundColor which lets the user enter a new RGBA for the window background and WindowSize which allows the user to edit the windows width and height however the method doesn't work in fullscreen and prints an error to the console instead.